pyproject.toml
[project]
name = "diffusion"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.7.0",
    "deepspeed>=0.18.7",
    "diffusers>=0.37.0",
    "einops>=0.8.2",
    "huggingface-hub>=1.5.0",
    "hydra-core>=1.3.2",
    "numpy>=2.4.2",
    "optimum>=2.1.0",
    "pytest>=9.0.2",
    "python-box>=7.4.1",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "timm>=1.0.25",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "wandb>=0.25.0",
    "weave>=0.52.30",
]
